requirements-dev.txt
# 开发和测试依赖

# 测试框架
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0  # 并行测试

# HTTP客户端
requests>=2.28.0
httpx>=0.24.0  # 异步HTTP客户端

# WebSocket客户端
websocket-client>=1.5.0

# 数据库
psycopg>=3.1.0  # PostgreSQL适配器
psycopg-binary>=3.1.0  # PostgreSQL适配器（二进制版本，Windows兼容）
redis>=4.5.0

# 代码质量
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# 类型注解
types-requests>=2.28.0
types-redis>=4.5.0

# 开发工具
ipython>=8.0.0
jupyter>=1.0.0

# 文档生成
sphinx>=5.0.0
sphinx-rtd-theme>=1.2.0

# 性能分析
memory-profiler>=0.60.0
line-profiler>=4.0.0

# 安全检查
bandit>=1.7.0
safety>=2.3.0

# Docker相关
docker>=6.0.0
docker-compose>=1.29.0
